Time functions in a separate module
