+++ backend/promis/util/orbit.py
@@ -23,13 +23,8 @@ import math
 import collections
 import operator
 import util.cubefit
-import datetime, pytz
 
 _earth_radius = 6371 # km
-
-# TODO: onboard time shift
-def maketime(u):
-    return datetime.datetime.fromtimestamp(u, tz=pytz.utc)
 
 def sign(x):
     """Returns 1 for non-negative arguments and -1 otherwise."""
